refactor some of the arg parsing for params
