+++ disdat/common.py
@@ -438,16 +438,62 @@ def parse_args_tags(args_tag, to='dict'):
     return tag_thing
 
 
-def parse_params(params):
+def parse_params(cls, params):
     """
+    Create a dictionary of str->str arguments to str->python objects deser'd by Luigi Parameters
+
     Input is the string "--arg value --arg2 value2"
-    Convert to dict {'arg':value,'arg2':value2}
 
-    :param params: from argparse
-    :return:   dict {'arg':value,'arg2':value2}
+    Convert to dict {'arg':str,'arg2':str2}
+
+    then
+
+    Convert to dict {'arg':luigi.Parameter.value,'arg2':luigi.Parameter.value2}
+
+    Args:
+        cls (type[disdat.pipe.PipeTask]):
+        params: from argparse
+
+    Returns:
+         dict {'arg':value,'arg2':value2}
     """
 
-    return {k.lstrip('--'): v for k, v in zip(params[::2], params[1::2])}
+    params_str_dict = {k.lstrip('--'): v for k, v in zip(params[::2], params[1::2])}
+
+    return convert_str_params(cls, params_str_dict)
+
+
+def convert_str_params(cls, params_str):
+    """
+    This is similar to Luigi.Task.from_str_params(cls, params_str)
+    But we don't create the class here, and we outer loop through our params (not the classes
+    params).  We just want to convert each of the params that are in the class and in this dictionary
+    into the deserialized form.
+
+    NOTE:  This is somewhat dangerous and could break if Luigi changes around
+    this code.  The alternative is to use Luigi.load_task() but then we have to ensure
+    all the input parameters are "strings" and we have to then put special code
+    inside of apply to know when to create a class normally, or create it from the CLI.
+
+    Parameters:
+        params_str (dict): dict of str->str.  param name -> value .
+    """
+    kwargs = {}
+
+    cls_params = {n: p for n, p in cls.get_params()}  # get_params() returns [ (name, param), ... ]
+
+    for param_name, param_str in params_str.items():
+        if param_name in cls_params:
+            param = cls_params[param_name]
+            if isinstance(param_str, list):
+                kwargs[param_name] = param._parse_list(param_str)
+            else:
+                kwargs[param_name] = param.parse(param_str)
+        else:
+            _logger.error("Parameter {} is not defined in class {}.".format(param_name, cls.__name__))
+            raise ValueError("Parameter {} is not defined in class {}.".format(param_name, cls.__name__))
+
+    return kwargs
 
 
 def get_local_file_path(url):
